Remove User Statistics and Merge Auth with User
