+++ backend/__tests__/integration/user.js
@@ -2,8 +2,7 @@ const mongoose = require('mongoose'); // Connects to mongodb
 const app = require('./../../server.js') // Link to your server file
 const Verification = require('./../../models/verification')
 const supertest = require('supertest');
-const UserAuth = require('../../models/userAuth.js');
-const User = require('../../models/userInfo.js');
+const User = require('../../models/user.js');
 const request = supertest(app)
 
 beforeAll(() => {
@@ -16,7 +15,7 @@ it('Signup Endpoint Test - Successful Signup', async done => {
   jest.setTimeout(30000);
   payload = {
     username: "jonsnow",
-    password: "WinterIsComing123", 
+    password: "WinterIsComing123",
     email: "Jon.Snow@stark.com",
     fullname: "Jon Snow",
     profilepic: "FAKE BASE64 ENCODED STRING"
@@ -27,11 +26,6 @@ it('Signup Endpoint Test - Successful Signup', async done => {
   expect(response.body.msg).toBe("Registration success! Please sign in")
 
   // remove the test user
-  await UserAuth.findOneAndDelete(
-    { "username": "jonsnow" },
-    { "sort": { "_id": -1 } }
-  )
-
   await User.findOneAndDelete(
     { "username": "jonsnow" },
     { "sort": { "_id": -1 } }
@@ -44,7 +38,7 @@ it('Signup Endpoint Test - Missing Username', async done => {
   // Sends POST Request to /signup endpoint
   jest.setTimeout(30000);
   payload = {
-    password: "WinterIsComing123", 
+    password: "WinterIsComing123",
     email: "Jon.Snow@stark.com",
     fullname: "Jon Snow",
     profilepic: "FAKE BASE64 ENCODED STRING"
@@ -79,7 +73,7 @@ it('Signup Endpoint Test - Missing Email', async done => {
   jest.setTimeout(30000);
   payload = {
     username: "jonsnow",
-    password: "WinterIsComing123", 
+    password: "WinterIsComing123",
     fullname: "Jon Snow",
     profilepic: "FAKE BASE64 ENCODED STRING"
   }
@@ -96,7 +90,7 @@ it('Signup Endpoint Test - Missing Fullname', async done => {
   jest.setTimeout(30000);
   payload = {
     username: "jonsnow",
-    password: "WinterIsComing123", 
+    password: "WinterIsComing123",
     email: "Jon.Snow@stark.com",
     profilepic: "FAKE BASE64 ENCODED STRING"
   }
@@ -113,7 +107,7 @@ it('Signup Endpoint Test - Invalid Username', async done => {
   jest.setTimeout(30000);
   payload = {
     username: "Jon Snow",
-    password: "WinterIsComing123", 
+    password: "WinterIsComing123",
     email: "Jon.Snow@stark.com",
     fullname: "Jon Snow",
     profilepic: "FAKE BASE64 ENCODED STRING"
@@ -131,7 +125,7 @@ it('Signup Endpoint Test - User already exists', async done => {
   jest.setTimeout(30000);
   payload = {
     username: "Cheng",
-    password: "Test123", 
+    password: "Test123",
     email: "test@test.com",
     fullname: "C L",
     profilepic: "FAKE BASE64 ENCODED STRING"
@@ -149,7 +143,7 @@ it('Signup Endpoint Test - Invalid Password', async done => {
   jest.setTimeout(30000);
   payload = {
     username: "jonsnow",
-    password: "WinterIsComing", 
+    password: "WinterIsComing",
     email: "Jon.Snow@stark.com",
     fullname: "Jon Snow",
     profilepic: "FAKE BASE64 ENCODED STRING"
@@ -167,7 +161,7 @@ it('Signup Endpoint Test - Invalid Email', async done => {
   jest.setTimeout(30000);
   payload = {
     username: "jonsnow",
-    password: "WinterIsComing123", 
+    password: "WinterIsComing123",
     email: "Jon.Snow@stark",
     fullname: "Jon Snow",
     profilepic: "FAKE BASE64 ENCODED STRING"
@@ -186,7 +180,7 @@ it('Login Endpoint Test - Successful Login', async done => {
   // Sends POST Request to /login endpoint
   jest.setTimeout(30000);
   payload = {
-    username : "Mitul2",
+    username : "aryastark",
     password : "Test123"
   }
   const response = await request.post('/api/user/login').send(payload)
@@ -405,7 +399,7 @@ it('Change Password Endpoint Test - Incorrect Verification Code', async done => 
 it('Change Password Endpoint Test - Successful Password Change', async done => {
   // Sends POST Request to /resetpassword endpoint
   jest.setTimeout(30000);
-  const username = "Mitul2"
+  const username = "aryastark"
 
   // send email
   payload = {
@@ -435,7 +429,7 @@ it('Change Password Endpoint Test - Successful Password Change', async done => {
 it('Change Password Endpoint Test - Block Re-use of verificaiton code', async done => {
   // Sends POST Request to /resetpassword endpoint
   jest.setTimeout(30000);
-  const username = "Mitul2"
+  const username = "aryastark"
 
   // send email
   payload = {
@@ -470,6 +464,125 @@ it('Change Password Endpoint Test - Block Re-use of verificaiton code', async do
   expect(response.status).toBe(200)
   expect(response.body.success).toBe("false")
   expect(response.body.msg).toBe("Verification Code is incorrect, expired or already used")
+  done()
+})
+
+it('Get User Information - No x-auth-username', async done => {
+  // Sends POST Request to /signup endpoint
+  jest.setTimeout(30000);
+
+  var expectedResponse = { "msg": "No username, authorization denied" }
+  const response = await request.get('/api/userprofile/userinformation/sansastark')
+
+  expect(response.body.msg).toBe(expectedResponse.msg);
+  done()
+})
+
+it('Get User Information - No x-auth-token', async done => {
+  // Sends POST Request to /signup endpoint
+  jest.setTimeout(30000);
+
+  var expectedResponse = { "msg": "No token, authorization denied" }
+  const response = await request.get('/api/userprofile/userinformation/sansastark').set('x-auth-username', 'sansastark')
+
+  expect(response.body.msg).toBe(expectedResponse.msg);
+  done()
+})
+
+it('Get User Information - Wrong x-auth-token', async done => {
+  // Sends POST Request to /signup endpoint
+  jest.setTimeout(30000);
+
+  var expectedResponse = { "msg": "Token is invalid" }
+  const response = await request.get('/api/userprofile/userinformation/sansastark').set('x-auth-username', 'sansastark').set('x-auth-token', "fake")
+
+  expect(response.body.msg).toBe(expectedResponse.msg);
+  done()
+})
+
+it('Get User Information - Wrong x-auth-username', async done => {
+  // Sends POST Request to /signup endpoint
+  jest.setTimeout(30000);
+
+  payload1 = {
+    username : "sansastark",
+    password : "WinterIsComing123"
+  }
+  const response2 = await request.post('/api/user/login').send(payload1)
+  expect(response2.status).toBe(200)
+  expect(response2.body.success).toBe("true")
+  expect(response2.body.msg).toBe("Login successful")
+
+  var token = response2.body.token
+  var expectedResponse = { "msg": "Token is invalid" }
+  const response = await request.get('/api/userprofile/userinformation/sansastark').set('x-auth-username', 'FAKEFAKE').set('x-auth-token', token)
+  expect(response.body.msg).toBe(expectedResponse.msg);
+
+  done()
+})
+
+
+it('Get User Information - Conflicting x-auth-username and userinformation/{username}', async done => {
+  // Sends POST Request to /signup endpoint
+  jest.setTimeout(30000);
+
+  payload1 = {
+    username : "sansastark",
+    password : "WinterIsComing123"
+  }
+  const response2 = await request.post('/api/user/login').send(payload1)
+  expect(response2.status).toBe(200)
+  expect(response2.body.success).toBe("true")
+  expect(response2.body.msg).toBe("Login successful")
+
+  var token = response2.body.token
+  var expectedResponse = {
+      "status": "400",
+      "success": "false",
+      "msg": "Bad Request"
+  }
+  const response = await request.get('/api/userprofile/userinformation/jonsnow').set('x-auth-username', 'sansastark').set('x-auth-token', token)
+  expect(response.body.msg).toBe(expectedResponse.msg);
+  expect(response.body.status).toBe(expectedResponse.status);
+  expect(response.body.success).toBe(expectedResponse.success);
+
+  done()
+})
+
+it('Get User Information - Successful', async done => {
+  // Sends POST Request to /signup endpoint
+  jest.setTimeout(30000);
+
+  payload1 = {
+    username : "sansastark",
+    password : "WinterIsComing123"
+  }
+  const response2 = await request.post('/api/user/login').send(payload1)
+  expect(response2.status).toBe(200)
+  expect(response2.body.success).toBe("true")
+  expect(response2.body.msg).toBe("Login successful")
+
+  var token = response2.body.token
+  var expectedResponse = {
+      "status": "200",
+      "success": "true",
+      "data": {
+          "username": "sansastark",
+          "fullname": "Jon Snow",
+          "email": "Jon.Snow@stark.com",
+          "totaldistancecomplete": 0,
+          "totalroutescomplete": 0,
+          "totalroutetime": "0",
+          "profilepic": "FAKE BASE64 ENCODED STRING"
+      },
+      "msg": "User information sent"
+  }
+  const response = await request.get('/api/userprofile/userinformation/sansastark').set('x-auth-username', 'sansastark').set('x-auth-token', token)
+
+  expect(response.body.status).toBe(expectedResponse.status);
+  expect(response.body.success).toBe(expectedResponse.success);
+  expect(response.body.data.username).toBe(expectedResponse.data.username);
+
   done()
 })
