+++ backend/__tests__/unit/models/user.js
@@ -1,6 +1,6 @@
 // Boilerplate https://bcostabatista.medium.com/testing-nodejs-applications-with-jest-7ae334daaf55
-const UserAuth = require('../../../models/userAuth')
-const {encrypt} = require('../../../controllers/userAuth.js')
+const User = require('../../../models/user')
+const {encrypt} = require('../../../controllers/user.js')
 const mongoose = require('mongoose'); // Connects to mongodb
 
 beforeAll(() => {
@@ -14,12 +14,12 @@ beforeAll(() => {
   console.error = function() {}
 });
 
-describe("User Authenticate", () => {
+describe("User", () => {
   test('Successfully Authenticate Username and Password', async () => {
     jest.setTimeout(30000);
-    username = "Mitul2"
+    username = "aryastark"
     password = "Test123"
-    const user = await UserAuth.findOne({ username })
+    const user = await User.findOne({ username })
     const encryptedPassword = encrypt(password)
     expect(user.authenticate(encryptedPassword)).toBe(true);
   });
@@ -29,7 +29,7 @@ describe("User Authenticate", () => {
     username = "Mitul2"
     oldpassword = "Test123"
     newpassword = "NewPassword123"
-    var user = await UserAuth.findOne({ username })
+    var user = await User.findOne({ username })
     var encryptedPassword = encrypt(newpassword)
     user.changepassword(encryptedPassword)
     expect(user.authenticate(encryptedPassword)).toBe(true);
